+++ src/ui/view.cpp
@@ -27,19 +27,19 @@ void View::do_resize(int width, int height)
     on_resize(width, height);
 }
 
-void View::do_mouse_press(MouseButton button, int x, int y)
+void View::do_mouse_press(MouseButton button, double x, double y)
 {
     on_mouse_press(button, x, y);
 }
-void View::do_mouse_release(MouseButton button, int x, int y)
+void View::do_mouse_release(MouseButton button, double x, double y)
 {
     on_mouse_release(button, x, y);
 }
-void View::do_mouse_moved(int x, int y)
+void View::do_mouse_moved(double x, double y)
 {
     on_mouse_moved(x, y);
 }
-void View::do_mouse_wheel(bool is_horizontal, float delta)
+void View::do_mouse_wheel(bool is_horizontal, double delta)
 {
     on_mouse_wheel(is_horizontal, delta);
 }
